pyproject.toml
[tool.poetry]
name = "loudreadingskill"
version = "2.0.0"
description = ""
authors = ["Silent <mikolaj@statystyka.net>"]
readme = "README.md"
packages = [
    { include = "client", from = "." },
    { include = "server", from = "." },
    { include = "core", from = "." },
]
include = ["data/audio/*.mp3"]

[tool.setuptools.packages.find]
where = "."

[tool.setuptools.package-data]
"data.audio" = ["*.mp3"]


[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pillow = "^10.3.0"
openai = "^1.30.3"
openai-whisper = { git = "https://github.com/openai/whisper.git", branch = "main", optional = true }
gtts = "^2.5.1"
sounddevice = "^0.4.7"
pyaudio = "^0.2.14"
pydub = "^0.25.1"
requests = "^2.32.3"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
numpy = "^1.19.0"
humanize = "^4.9.0"
overrides = "^7.7.0"
pynvml = { version = "^11.0.0", optional = true }
coqui-tts = { version = "^0.24", optional = true }
torch = { version = "^2.0", optional = true }
torchaudio = { version = "^2.0.0", optional = true }

[tool.poetry.extras]
local = ["openai-whisper", "pynvml"]
respeak = ["coqui-tts", "torch", "torchaudio"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
loudreading_server = "server.whisper_server:init"
loudreading_client = "client.reading:main"
